tools/xdp-forwarder: Do not include libc headers in eBPF programs
The build happened to work on arm64 because the glibc arm64 headers
don't support multilib.  On x86_64, glibc headers assume that BPF is a
32-bit platform (because __x86_64__ isn't defined) and fail to find the
32-bit headers.  This is not a glibc bug.  Rather, BPF programs should
not be including glibc headers.

Most Linux headers are not trivial to include in BPF programs.  The
version of the headers meant for userspace use do include glibc headers,
and that isn't supported in BPF.  The version meant for building
kernel modules does not, but using it requires much more complicated
build system.

Solve this problem by only including headers intended for use in BPF
programs.  These headers include declarations explicitly intended for
use in BPF programs, so if they do pull in libc headers that is a bug.
Nix's wrapped clang would pull in libc headers automatically and is not
suitable when a target is specified explicitly.  Therefore, use an
unwrapped clang.

Signed-off-by: Demi Marie Obenour <demiobenour@gmail.com>
---
To check that the program does not include libc headers, one can add
#include <stdio.h> (or any other libc header) and check that one gets
an appropriate error.

Signed-off-by: Demi Marie Obenour <demiobenour@gmail.com>