host/rootfs: Install systemd-pull
Due to a systemd bug [1], building systemd-sysupdate does not require
that systemd-pull is built as well.  However, systemd-sysupdate has a
run-time dependency on systemd-pull.  Therefore, override the systemd
derivation so that systemd-pull is built.  Confusingly, this requires
enabling systemd-importd.

If systemd-pull or systemd-sysupdate is not built, the resulting image
will be broken and users will not be able to recover without either a
reinstall or reverting to the previous version.  Therefore, add a check
to ensure that both are in fact built.  Use 'cat' rather than just
'stat' to catch broken symlinks and the like.

Signed-off-by: Demi Marie Obenour <demiobenour@gmail.com>