host/rootfs: /etc/mdev/net/add: make idempotent
udev actions needs to be idempotent, so when we switch to udev, we
have the choice of making this script idempotent, or doing goto hacks
to stop udev running it more than once.  I like this better.

assign-devices is already idempotent, so all we need to do to achieve
idempotency is avoid unbinding and rebinding the driver if it's
already the right one.

Link: https://spectrum-os.org/lists/archives/spectrum-devel/87ikgxwq0v.fsf@alyssa.is
Signed-off-by: Alyssa Ross <hi@alyssa.is>
Message-ID: <20251102150018.1242022-2-hi@alyssa.is>
Reviewed-by: Demi Marie Obenour <demiobenour@gmail.com>