Move verity and EFI creation to separate Nix derivations
This doesn't have any functional change, other than to use the read
builtin instead of a cat command in a shell script.  However, it does
make the code much cleaner and more reusable.  For instance, one can
easily build just the verity image or just the UKI.

This will be used by the Nix code that generates an update package.  The
update package needs the root filesystem, the verity superblock, and the
UKI.  It doesn't need the installer or the live image.

Signed-off-by: Demi Marie Obenour <demiobenour@gmail.com>

---
Changes in v5:
- Rebase
- Link to v4: https://spectrum-os.org/lists/archives/spectrum-devel/20251119-refactor-verity-v4-0-9bc56d5216c0@gmail.com

Changes in v4:
- Many cleanups.
- Respond to suggestions from code review.
- Link to v3: https://spectrum-os.org/lists/archives/spectrum-devel/20251111-refactor-verity-v3-0-575726639f9e@gmail.com

Changes in v3:
- Rebase on main
- Link to v2: https://spectrum-os.org/lists/archives/spectrum-devel/20251107-refactor-verity-v2-0-2af58b1a4a87@gmail.com

Changes in v2:
- Do not break interactive rootfs development.
- Link to v1: https://spectrum-os.org/lists/archives/spectrum-devel/20251105-refactor-verity-v1-0-b8ba27dfdf06@gmail.com

--- b4-submit-tracking ---
# This section is used internally by b4 prep for tracking purposes.
{
  "series": {
    "revision": 5,
    "change-id": "20251105-refactor-verity-9c8ca37e021a",
    "prefixes": [],
    "history": {
      "v1": [
        "20251105-refactor-verity-v1-0-b8ba27dfdf06@gmail.com"
      ],
      "v2": [
        "20251107-refactor-verity-v2-0-2af58b1a4a87@gmail.com"
      ],
      "v3": [
        "20251111-refactor-verity-v3-0-575726639f9e@gmail.com"
      ],
      "v4": [
        "20251119-refactor-verity-v4-0-9bc56d5216c0@gmail.com"
      ]
    }
  }
}