Validate configuration parameters
Wrong values for the version or update URL will cause very confusing
build-time or runtime errors.  Provide a better user experience by
validating them up-front.

The update URL validator is strict.  It rejects the following:

- Fragment specifiers.  This breaks appending /something to the URL.
- Query parameters.  This also breaks appending /something to the URL.
- Non-ASCII domain names.  Musl getaddrinfo() does not support Unicode.
- Non-unreserved characters in paths that are not %-encoded.  The only
  correct use of these is matrix parameters, and I have only seen these
  used in exploits.
- Credentials (usernames and/or passwords).  These do not belong in an
  image meant for distribution.
- IPv6 interface specifiers.  Spectrum does not guarantee anything about
  the name of the UpdateVM's outgoing interface.
- Schemes other than http:// or https://.  These are hardly ever used
  for update distribution in 2025.

The version validator is in lib/config.nix, as the version number is
used in many places.  It checks that the version is of the form
MAJOR.MINOR.PATCH with an optional suffix.  It also checks that it
conforms to systemd's version number specification [1] and does not
violate a SHOULD NOT in that specification.

[1]: https://uapi-group.org/specifications/specs/version_format_specification

Signed-off-by: Demi Marie Obenour <demiobenour@gmail.com>