tools: start-vmm: enable Cloud Hypervisor landlock
We can't really predict the device paths or IOMMU groups statically,
so this is as good as it gets with landlock rules.  We'll be able to
do other things to further lock things down though, like running
different Cloud Hypervisor instances as different users, and changing
ownership of each IOMMU group in /dev/vfio/vfio to match.

Message-ID: <20251110184907.346511-1-hi@alyssa.is>
Signed-off-by: Alyssa Ross <hi@alyssa.is>