img/app: dbus: don't listen on VSOCK
After working on it for a while, I decided that it complicated the
D-Bus security model too much to upstream VSOCK support for the bus.
Proxying D-Bus with socat will allow us to drop the D-Bus VSOCK
patches.

The new dbus-vsock service starts before dbus-daemon to ensure that
VSOCK connections can be received as soon as
org.freedesktop.impl.portal.desktop.spectrum is started.  When a
connection is received (which should only be after the bus is up and
has started org.freedesktop.impl.portal.desktop.spectrum), it will be
relayed to the bus.

Sadly we do still need to allow ANONYMOUS authentication for now[1].

Link: https://github.com/z-galaxy/zbus/issues/1003#issuecomment-3523214990 [1]
Signed-off-by: Alyssa Ross <hi@alyssa.is>
Message-ID: <20251113111038.39098-2-hi@alyssa.is>
Acked-by: Demi Marie Obenour <demiobenour@gmail.com>