release: Compress installation images and remove live image
mkfs.ext4 is not able to produce images with files large enough to hold
both the primary and backup copy of the root partition [1].  Reducing
the sizes of partitions to be little greater than the size of the root
filesystem image does not help.  The produced file is still too large.
Therefore, compress the image, which causes it to be small enough that
mkfs.ext4 can handle it.  This breaks the live image, so remove it.
The live image will return once Spectrum switches to the GNOME OS
installer [2].

[1]: https://github.com/tytso/e2fsprogs/issues/254
[2]: https://spectrum-os.org/lists/archives/spectrum-devel/87wm4dlkhz.fsf@alyssa.is/

Signed-off-by: Demi Marie Obenour <demiobenour@gmail.com>