img/app: remove XDG_RUNTIME_DIR
Nothing uses this any more, and let's hope it stays that way, because
it makes it difficult to run services as different users.  Maybe we
end up finding something that really needs it in the future, but it
makes things much easier to avoid it for as long as we can.

Older applications might not support WAYLAND_DISPLAY being absolute,
but this has been changed in libwayland and GTK.  We don't need to
support arbitrarily old applications, and this problem will resolve
itself over time.

Signed-off-by: Alyssa Ross <hi@alyssa.is>
Message-ID: <20251126213407.1773744-6-hi@alyssa.is>