host/rootfs: stop using pkgsStatic
For as long as we're running mesa on the host, we'll need dynamic
linking, so let's at least stick to a single package set.

Since we're longer cross compiling, I've switched to taking inkscape
and xorg from the default package set.  They're only used for
building, not installed in the image, so let's avoid having to build
their large musl closures.

Signed-off-by: Alyssa Ross <hi@alyssa.is>
Message-ID: <20251127164119.2037885-3-hi@alyssa.is>