scripts/make-gpt.sh: allow setting partition size
We want non-minimally-sized partitions to leave space for updates.
All such partitions will also be labelled, so we can just add another
optional field at the end.

Since we don't parse partition specifications in sh, we can't keep a
running total any more, so instead we just go through the table at the
end and add up all the sizes, taking advantage of our knowledge that
the size will always be the last thing in each line in our tables.

Signed-off-by: Alyssa Ross <hi@alyssa.is>
Message-ID: <20251128214406.465618-2-hi@alyssa.is>