release: Compress installation images and remove live image
systemd-sysupdate will fail if the OS image does not fit in the
partitions that the installer created.  Therefor, the partitions need to
be very large so that there is plenty of room for the OS to grow.
Furthermore, systemd-sysupdate requires both A and B copies of both the
root and verity partitions.

mkfs.ext4 is not able to produce images with files large enough to hold
both the primary and backup copy of the root partition [1].  Reducing
the sizes of partitions to be little greater than the size of the root
filesystem image does not help.  The produced file is still too large.
Therefore, compress the image, which causes it to be small enough that
mkfs.ext4 can handle it.

This breaks the option to use the installer as a live image.  Therefore,
remove it.  This option will return once Spectrum switches to the GNOME
OS installer [2].  However, it is still possible to build a live image
that is separate from the installer.  Document how to build and use it.

GRUB2 does support compressed loopback images, but these presumably
buffer the whole image in memory.  Since the entire installer will be
replaced, making it work is not considered worthwhile.

[1]: https://github.com/tytso/e2fsprogs/issues/254
[2]: https://spectrum-os.org/lists/archives/spectrum-devel/87wm4dlkhz.fsf@alyssa.is/

Signed-off-by: Demi Marie Obenour <demiobenour@gmail.com>
Reviewed-by: Alyssa Ross <hi@alyssa.is>
Message-ID: <20251129-updates-v6-2-9edb87a2e509@gmail.com>
Signed-off-by: Alyssa Ross <hi@alyssa.is>