Use OS version to set partition labels and UKI name
systemd-sysupdate has strict requirements on the partition layout:

- The label of the active partition must match the template in the
  .transfer file.  For instance, the root filesystem of Spectrum 0.0.0
  must be in a partition with label "Spectrum_0.0.0", and the verity
  partition must have the label "Spectrum_0.0.0.verity".

- The label of the inactive partition must be that of the old version of
  Spectrum, or "_empty" for freshly installed systems.

- The partition type UUID must conform to the Discoverable Partition
  Specification.

Also, the UKI must have a name that includes the OS version.  Otherwise,
it will not be deleted during updates.

Since the partition label includes the OS version, add an OS version
number.  Use 0.0.0 to indicate that Spectrum OS is still in very early
development and should not be used.  The version number can be
overridden in the build configuration file.

Signed-off-by: Demi Marie Obenour <demiobenour@gmail.com>
Reviewed-by: Alyssa Ross <hi@alyssa.is>
Message-ID: <20251129-updates-v6-3-9edb87a2e509@gmail.com>
Signed-off-by: Alyssa Ross <hi@alyssa.is>