release/checks/integration: handle partial sends
Sometimes we see a test failure that looks like this:

	unexpected connection data: qemu-system-aarch64: terminating on signal 15 from pid 184 ()

I think this is because, now that the nc command in the networking
integration test has a timeout, it's possible for it to time out after
having opened the connection, but before having written all its input
to it.  Therefore, ignore connections that send a prefix of the
expected data (including nothing), and just wait for the next
connection rather than failing if that happens.

Closes: https://spectrum-os.org/lists/archives/spectrum-devel/875xbl33vw.fsf@alyssa.is/
Fixes: c61b297 ("release/checks/integration: add nc timeout")
Signed-off-by: Alyssa Ross <hi@alyssa.is>
Message-ID: <20251129180016.3895-2-hi@alyssa.is>
Reviewed-by: Yureka Lilian <yureka@cyberchaos.dev>