host/rootfs: add run-flatpak script
This is the entrypoint for running Flatpak applications.

It would be good to only add mounts for the VM in virtiofsd's mount
namespace, so we don't need to do lots of manual unmounts, but that's
a wider change affecting more than just Flatpak.

I've tested this by copying my host's Flatpak repository into a disk
image, and attaching that as a drive to the VM.

Signed-off-by: Alyssa Ross <hi@alyssa.is>
Message-ID: <20251201170458.4186683-11-hi@alyssa.is>