host/rootfs: fix portal writes
Now that applications in the guest run as uid/gid 1000, we need
virtiofsd to pass through files as owned by those same uids/gids.
virtiofsd's built-in remapping feature doesn't work when running as
root, so we do it manually.  This is likely only a temporary measure
anyway until we have virtiofsd running as non-root on the host.

Fixes: 8bfcbf9 ("img/app: run applications as non-root")
Signed-off-by: Alyssa Ross <hi@alyssa.is>