img/app: fix running AppImages
The non-root user that applications run as can't mount FUSE
filesystems.  I don't want to introduce setuid or an IPC service just
for running AppImages, although we could revisit if we find that
applications themselves rely on being able to mount FUSE filesystems.
For now, it'll be more efficient to use the in-kernel squashfs
implementation anyway.

We do need to set a few environment variables normally set by the
AppImage runtime, but as far as I can tell it doesn't do much more
than this that applications could rely on.

This only works for type 2 AppImages, but I expect that's just about
every AppImage nowadays, and if turns out not to be it shouldn't be
too hard to support type 1.

Fixes: 8bfcbf9 ("img/app: run applications as non-root")
Signed-off-by: Alyssa Ross <hi@alyssa.is>
Message-ID: <20251208211101.430776-1-hi@alyssa.is>