img/app: fix X11
Xwayland only accepts connections from the user it's running as.  It
is started by wayland-proxy-virtwl, which does not allow passing extra
options, so we can't change its authentication method.

Therefore, the only way for X11 to work with the current software is
to run wayland-proxy-virtwl as the same user as the application.

I expect that in the near future, we will use xwayland-satellite
instead of the built-in Xwayland translation in wayland-proxy-virtwl.
When that happens, we can run the stub compositor as its own user
again.

Fixes: cb27e3a ("img/app: wayland-proxy-virtwl: run as non-root")
Signed-off-by: Alyssa Ross <hi@alyssa.is>
Message-ID: <20251208211607.434907-1-hi@alyssa.is>