tools/router: use socket activation
This means we can use readiness notification to wait until the sockets
are created without having to add special functionality for that to
the router program, and also means we can do extra system-specific
setup to the sockets, like changing their owners, outside of the
router.

Since the socket paths were the only arguments taken by the router,
this also lets us drop the clap dependency entirely.

Signed-off-by: Alyssa Ross <hi@alyssa.is>
Message-ID: <20251209125442.719952-2-hi@alyssa.is>