host/rootfs: add root-terminal s6-sudod service
This will be needed as an interim measure to allow Weston itself to be
run as non-root, without the user losing access to root entirely in
the graphical environment.  Currently, only root can access this
service, so it's redundant, but we can later make the socket
accessible to Weston's user.

Signed-off-by: Alyssa Ross <hi@alyssa.is>
Message-ID: <20251209182402.872822-1-hi@alyssa.is>