host/rootfs: add mount-userdata command
We shouldn't leave it to the user to manually mount user data
partitions.  We want to ensure restrictive mount flags are set, and be
able to do other operations around the mounting.  It's also nice to
have a single place where all user data partitions can be mounted.
I've taken inspiration from the FHS in using /media for this.

Signed-off-by: Alyssa Ross <hi@alyssa.is>
Message-ID: <20251212214859.389183-1-hi@alyssa.is>