host/roots: Sandbox xdg-desktop-portal-spectrum-host
It is quite possible that these Landlock rules are unnecessarily
permissive, but all of the paths to which read and execute access is
granted are part of the root filesystem and therefore assumed to be
public knowledge.  Removing access from any of them would only increase
the risk of accidental breakage in the future, and would not provide any
security improvements.  seccomp *could* provide some improvements, but
the effort needed is too high for now.

Signed-off-by: Demi Marie Obenour <demiobenour@gmail.com>