host/rootfs: give VMs a disk-backed directory
This is the first step towards persistent VM data.  For now, these
directories are never reused, and also don't get cleaned up.  Both are
things to work on in future.  Even without persistence, it's good to
not have to store everything a VM might write into its home directory
in RAM.

For AppImage and Flatpak VMs, disk-backed directories are stored on
the partition containing the AppImage or Flatpak, in the new Spectrum/
hierarchy.  This will enable locating them for later reuse.  System
VMs, on the other hand, don't have a natural partition to store data
on — there may not even be a writable partition at the time they're
launched.  Since we don't expect persistence from system VMs, we just
use a tmpfs for their "disk-backed" directories, which is reset on
every boot.  Doing it this way avoids the need for multiple routes in
img/app — it can always assume it gets a "disk" directory from the
host.  For manually imported VMs, it's up to the user where these
directories should be, or whether they should have the same behavior
as system VMs.

Signed-off-by: Alyssa Ross <hi@alyssa.is>
Message-ID: <20251214014229.775825-4-hi@alyssa.is>