tools/vm-set-persist.c: init
This allows the disk-backed directory of a running VM to be made
persistent, with a user-provided name.  This is done by renaming the
directory to have a "persist." prefix rather than the "tmp." one the
cleaner will look for.  Since the VM's virtiofsd accesses the
directory via a bind mount, this rename will be unnoticeable to the
guest.

musl has quite a bit of catching up to do with the APIs used here,
which requires the use of a lot of raw syscalls.  This even applies to
some syscalls musl has wrappers for, like mkdirat(2), because musl's
mkdirat() comes from <sys/stat.h>, which defines a struct statx that's
missing the stx_mnt_id member we need.  There doesn't even seem to be
a SYS_statmount, so we use __NR_ constants throughout for consistency.

Signed-off-by: Alyssa Ross <hi@alyssa.is>
Message-ID: <20251214014229.775825-10-hi@alyssa.is>