release/checks/integration: fix race condition
By default, IPv6 addresses are in the tentative state and can't be
bound to until the kernel has asychronously done Duplicate Address
Detection.  This was the cause of us needing to retry the bind, and
sometimes even that wasn't enough.  Since we're in a fresh network
namespace, we know there can't be a duplicate address, so we can add
the address with the IFA_F_NODAD flag to skip this and make the
address available for binding immediately.  We can't set that flag
using the ioctl interface we were using before, so switch to netlink.

This is slightly lazy netlink — we hardcode the index of the loopback
address, and don't bother with extended errors — but I think that's
fine for a test where readability is the most important thing.

I've run the test 100 times with this change and didn't see any bind
failures, which is a big improvement over the failure rate I was
seeing before.

Fixes: cdf1891e ("release/checks/integration: Adapt networking test for ipv6")
Signed-off-by: Alyssa Ross <hi@alyssa.is>
Message-ID: <20251218185830.89957-1-hi@alyssa.is>